backend/requirements.txt
# Core LangChain and LangGraph
langchain>=0.3.0
langgraph>=0.1.0
langchain-google-genai>=0.1.0
langchain-community>=0.0.20
langchain-tavily>=0.1.0

# Web Framework and API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# CLI Interface
typer[all]>=0.9.0
rich>=13.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Search and Web Scraping
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development and Monitoring
langsmith>=0.1.0
langchain-core>=0.1.0
httpx>=0.25.0

# Utilities
python-multipart>=0.0.6

# Development tools
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
bandit>=1.7.0
safety>=2.3.0 